moving RectorApplication to PhpFileProcessor
