+++ tests/Application/ApplicationFileProcessor/config/configured_rule.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-use Rector\Core\Tests\Application\ApplicationFileProcessor\Source\TextNonPhpFileProcessor;
+use Rector\Core\Tests\Application\ApplicationFileProcessor\Source\TextFileProcessor;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
-    $services->set(TextNonPhpFileProcessor::class);
+    $services->set(TextFileProcessor::class);
 };
