+++ utils/phpstan-extensions/src/TypeAnalyzer/AllowedAutoloadedTypeAnalyzer.php
@@ -18,7 +18,7 @@ final class AllowedAutoloadedTypeAnalyzer
      * @see https://regex101.com/r/BBm9bf/1
      * @var string
      */
-    private const AUTOLOADED_CLASS_PREFIX_REGEX = '#^(PhpParser|PHPStan|Rector|Reflection)#';
+    private const AUTOLOADED_CLASS_PREFIX_REGEX = '#^(PhpParser|PHPStan|Rector|Reflection|Symfony\\\\Component\\\\Console)#';
 
     /**
      * @var array<class-string>
