+++ utils/phpstan-extensions/tests/Rule/NoInstanceOfStaticReflectionRule/Fixture/InstanceofWithType.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Rector\PHPStanExtensions\Tests\Rule\NoInstanceOfStaticReflectionRule\Fixture;
 
-use Symfony\Component\Console\Command\Command;
+use Hoa\Math\Sampler\Random;
 
 final class InstanceofWithType
 {
     public function check($object)
     {
-        if ($object instanceof Command) {
+        if ($object instanceof Random) {
             return true;
         }
     }
