+++ utils/phpstan-extensions/tests/Rule/NoInstanceOfStaticReflectionRule/Fixture/IsAWithType.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Rector\PHPStanExtensions\Tests\Rule\NoInstanceOfStaticReflectionRule\Fixture;
 
-use Symfony\Component\Console\Command\Command;
+use Hoa\Math\Sampler\Random;
 
 final class IsAWithType
 {
     public function check($object)
     {
-        return is_a($object, Command::class);
+        return is_a($object, Random::class);
     }
 }
