+++ utils/phpstan-extensions/tests/Rule/NoInstanceOfStaticReflectionRule/NoInstanceOfStaticReflectionRuleTest.php
@@ -37,6 +37,7 @@ final class NoInstanceOfStaticReflectionRuleTest extends AbstractServiceAwareRul
         yield [__DIR__ . '/Fixture/SkipReflection.php', []];
         yield [__DIR__ . '/Fixture/SkipDateTime.php', []];
         yield [__DIR__ . '/Fixture/SkipTypesArray.php', []];
+        yield [__DIR__ . '/Fixture/SkipSymfony.php', []];
     }
 
     protected function getRule(): Rule
